Add form validation to the accessory request form
Create a backend AccessoryRequestForm to feed the data from the frontend form in order to use Django's is_valid() for the form validation. Add the divs for the error messages in the html template and rename the "name" property values to match those in the AccessoryRequest model. Also add custom styling for the error messages, as apparently they would display without any colours due to Bootstrap not having an alert-error class.